backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# LLM Integration
openai==1.3.7
tiktoken==0.5.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Utilities
structlog==23.2.0
python-json-logger==2.0.7
click==8.1.7
rich==13.7.0

# Monitoring & Metrics
psutil==5.9.6
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0
factory-boy==3.3.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings[python]==0.24.0
